Refactor path implementations to extend ReadOnlyPath
- Introduced a new abstract class ReadOnlyPath to encapsulate common behavior for remote, read-only file systems.
- Updated CmisPath, HdfsPath, S3Path, SmbPath, and WebDavPath to extend ReadOnlyPath, reducing code duplication.
- Removed redundant path parsing logic from individual path classes, leveraging the new parse method in ReadOnlyPath.
- Simplified path normalization and segment handling by utilizing methods from ReadOnlyPath.
- Ensured all path classes maintain their absolute path characteristics while adhering to the new structure.
